move chan_close_init
